Made a global styles sheet and used it, refactor RenderField into its own folder and refactored imports
