+++ client/src/styles/global.css
@@ -1,0 +1,4 @@
+/* Margin 0 */
+:global(.mb-0) {
+    margin-bottom: 0px;
+}
